Change project structure for packaging, add pyproject.toml
